requirements.txt
# Core dependencies
langgraph>=0.0.65
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0

# Data processing
pandas>=2.0.0
datasets>=2.14.0
jsonlines>=4.0.0

# Fine-tuning
transformers>=4.35.0
torch>=2.1.0
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.24.0
trl>=0.7.0

# Evaluation
scikit-learn>=1.3.0
nltk>=3.8.0
rouge-score>=0.1.2

# Configuration and utilities
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0
tqdm>=4.66.0
colorama>=0.4.6

# Development
pytest>=7.4.0
black>=23.0.0
ruff>=0.1.0
